feat: refactor module imports to use lowercase naming convention
